+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.4.30] - 2025-07-28
+
+### Changed
+
+- Style sheet improvements.
+
+### Fixed
+
+- Ruler efficiency and fixes.
+
 ## [0.4.29] - 2025-07-27
 
 ### Fixed